requirements.txt
eth-brownie==1.18.2
dotmap==1.3.8
pycoingecko==2.2.0
discord
discord.py
python-dotenv